requirements.txt
flask==2.0.1
flask-restful==0.3.9
playwright==1.32.0
beautifulsoup4==4.10.0
requests==2.28.1
pandas==1.4.2
openai==0.27.0
apscheduler==3.9.1
python-dotenv==0.20.0
lxml==4.9.1
tqdm==4.64.0
bs4==0.0.1
